# production_backend/docs/openapi.generated.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PumpDeviceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PumpDeviceRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PumpDeviceListResponse",
        "type": "object"
      },
      "PumpDeviceRead": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "firmware_version": {
            "title": "Firmware Version",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "device_id",
          "model",
          "firmware_version",
          "status"
        ],
        "title": "PumpDeviceRead",
        "type": "object"
      },
      "PumpDeviceUpsert": {
        "properties": {
          "firmware_version": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Version"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "title": "PumpDeviceUpsert",
        "type": "object"
      },
      "PumpEnergyTargetRead": {
        "properties": {
          "lower_value": {
            "title": "Lower Value",
            "type": "integer"
          },
          "upper_value": {
            "title": "Upper Value",
            "type": "integer"
          }
        },
        "required": [
          "lower_value",
          "upper_value"
        ],
        "title": "PumpEnergyTargetRead",
        "type": "object"
      },
...
      "PumpTelemetryEventCreate": {
        "properties": {
          "device_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "event_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Event Type",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "device_id",
          "event_type",
          "occurred_at"
        ],
        "title": "PumpTelemetryEventCreate",
        "type": "object"
      },
      "PumpTelemetryEventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PumpTelemetryEventRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PumpTelemetryEventListResponse",
        "type": "object"
      },
      "PumpTelemetryEventRead": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "device_id",
          "event_type",
          "occurred_at",
          "payload"
        ],
        "title": "PumpTelemetryEventRead",
        "type": "object"
      },
      "PumpThresholdCreate": {
        "properties": {
          "deep_level_l": {
            "exclusiveMinimum": 0.0,
            "title": "Deep Level L",
            "type": "integer"
          },
          "deep_level_r": {
            "exclusiveMinimum": 0.0,
            "title": "Deep Level R",
            "type": "integer"
          },
          "device_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "device",
            "title": "Source"
          },
          "stimulate_level_l": {
            "exclusiveMinimum": 0.0,
            "title": "Stimulate Level L",
            "type": "integer"
          },
          "stimulate_level_r": {
            "exclusiveMinimum": 0.0,
            "title": "Stimulate Level R",
            "type": "integer"
          }
        },
        "required": [
          "device_id",
          "occurred_at",
          "stimulate_level_l",
          "deep_level_l",
          "stimulate_level_r",
          "deep_level_r"
        ],
        "title": "PumpThresholdCreate",
        "type": "object"
      },
      "PumpThresholdRead": {
        "properties": {
          "deep_level_l": {
            "title": "Deep Level L",
            "type": "integer"
          },
          "deep_level_r": {
            "title": "Deep Level R",
            "type": "integer"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stimulate_level_l": {
            "title": "Stimulate Level L",
            "type": "integer"
          },
          "stimulate_level_r": {
            "title": "Stimulate Level R",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "device_id",
          "occurred_at",
          "stimulate_level_l",
          "deep_level_l",
          "stimulate_level_r",
          "deep_level_r",
          "source"
        ],
        "title": "PumpThresholdRead",
        "type": "object"
      },
      "PumpWorkstateCreate": {
        "properties": {
          "device_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "device",
            "title": "Source"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          }
        },
        "required": [
          "device_id",
          "occurred_at"
        ],
        "title": "PumpWorkstateCreate",
        "type": "object"
      },
      "PumpWorkstateRead": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "device_id",
          "occurred_at",
          "source",
          "state"
        ],
        "title": "PumpWorkstateRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MomCozy Production Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/devices/pump-energy-target": {
      "get": {
        "operationId": "get_pump_energy_target_v1_devices_pump_energy_target_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpEnergyTargetRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Pump Energy Target",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
...
    "/v1/devices/pump-telemetry": {
      "get": {
        "operationId": "list_pump_telemetry_v1_devices_pump_telemetry_get",
        "parameters": [
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpTelemetryEventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pump Telemetry",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      },
      "post": {
        "operationId": "create_pump_telemetry_v1_devices_pump_telemetry_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PumpTelemetryEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpTelemetryEventRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Pump Telemetry",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/devices/pump-threshold": {
      "post": {
        "operationId": "create_pump_threshold_v1_devices_pump_threshold_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PumpThresholdCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpThresholdRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Pump Threshold",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/devices/pump-threshold/latest": {
      "get": {
        "operationId": "get_latest_pump_threshold_v1_devices_pump_threshold_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "device_id",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpThresholdRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Latest Pump Threshold",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/devices/pump-workstate": {
      "post": {
        "operationId": "create_pump_workstate_v1_devices_pump_workstate_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PumpWorkstateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpWorkstateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Pump Workstate",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/devices/pump-workstate/latest": {
      "get": {
        "operationId": "get_latest_pump_workstate_v1_devices_pump_workstate_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "device_id",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpWorkstateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Latest Pump Workstate",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/devices/pumps": {
      "get": {
        "operationId": "list_pump_devices_v1_devices_pumps_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpDeviceListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pump Devices",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
    "/v1/devices/pumps/{device_id}": {
      "put": {
        "operationId": "upsert_pump_device_v1_devices_pumps__device_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PumpDeviceUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpDeviceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Pump Device",
        "tags": [
          "devices"
        ],
        "x-momcozy-api-surface": "public_app_api",
        "x-momcozy-client": [
          "flutter"
        ],
        "x-momcozy-owner": "devices",
        "x-momcozy-stability": "stable"
      }
    },
...
  }
}